requirements.txt
absl-py==0.11.0
addict==2.4.0
albumentations==1.3.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
anyio==3.6.2
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
attrdict==2.0.1
attrs==19.3.0
Automat==0.8.0
Babel==2.11.0
backcall==0.1.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
black==23.1.0
bleach==3.1.1
blessed==1.19.1
blinker==1.4
boto==2.49.0
bs4==0.0.1
cachetools==4.0.0
caffe==1.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
chumpy==0.70
class-resolver==0.4.2
click==8.1.3
codecov==2.1.12
colorama==0.4.6
command-not-found==0.3
commonmark==0.9.1
constantly==15.1.0
contourpy==1.0.7
coverage==7.0.1
crcmod==1.7
cryptography==2.8
cssselect==1.2.0
cssutils==2.6.0
cycler==0.10.0
Cython==0.29.32
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
descartes==1.1.0
dill==0.3.6
distlib==0.3.7
distro==1.4.0
distro-info==1.0
e2cnn==0.2.2
easyocr==1.7.1
einops==0.2.0
engineering-notation==0.7.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fair-esm==2.0.0
fastjsonschema==2.16.3
filelock==3.12.4
fire==0.5.0
flake8==3.7.9
Flask==2.1.3
Flask-Babel==2.0.0
flatbuffers==1.12
flyingcircus==0.1.4.0
fonttools==4.38.0
fsspec==2023.5.0
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
gdown==3.11.0
google-auth==1.5.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpg===1.13.1-unknown
grpcio==1.32.0
h11==0.14.0
h5py==2.10.0
html5lib==1.0.1
httpcore==0.18.0
httplib2==0.14.0
httpx==0.25.0
huggingface-hub==0.14.1
hyperlink==19.0.0
idna==2.8
imageio==2.4.1
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
incremental==16.10.1
iniconfig==1.1.1
install==1.3.5
interrogate==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==8.0.5
isort==4.3.21
itsdangerous==2.1.2
jedi==0.15.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==0.14.0
json-tricks==3.17.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-events==0.6.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lanms-neo==1.0.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Levenshtein==0.23.0
libclang==16.0.6
llvmlite==0.36.0
lmdb==1.0.0
lxml==4.9.2
lyft-dataset-sdk==0.0.8
Mako==1.1.0
Markdown==3.1.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matlab==0.1
matplotlib==3.3.2
mccabe==0.6.1
mdurl==0.1.2
mistune==2.0.5
mmaction2==1.1.0
-e git+https://github.com/open-mmlab/mmclassification.git@fb16bdc6a25f96843f8fb97b02e45872c61dc566#egg=mmcls
-e git+https://github.com/open-mmlab/mmdetection@f78af7785ada87f1ced75a2313746e4ba3149760#egg=mmdet
-e git+https://github.com/open-mmlab/mmdetection3d.git@47285b3f1e9dba358e98fcd12e523cfd0769c876#egg=mmdet3d
mmengine==0.7.3
mmocr==0.6.3
mmpose==1.1.0
mmsegmentation==0.30.0
model-index==0.1.11
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.2.1
multiprocess==0.70.14
munkres==1.1.4
mypy-extensions==1.0.0
mysqlclient==1.4.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.0.4
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==2.4
ninja==1.11.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.53.0
numpy==1.23.4
nvidia-ml-py==7.352.0
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.6.0.66
opencv-python-headless==4.2.0.34
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.0.10
openxlab==0.0.15
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
package-name==0.1
packaging==20.3
paddle-bfloat==0.1.7
paddleocr==2.6.1.2
paddlepaddle==2.5.1
paddlepaddle-gpu==2.5.1
pandas==2.0.1
pandocfilters==1.5.0
parsel==1.5.2
parso==0.5.2
pathspec==0.11.1
pdf2docx==0.5.6
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.2.0
pip-search==0.0.12
pipenv==2023.10.20
platformdirs==3.11.0
plotly==5.13.1
pluggy==1.0.0
plyfile==0.8.1
Polygon3==3.0.9.1
prefetch-generator==1.0.1
premailer==3.10.0
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==2.0.10
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.16.0
pycuda==2019.1.2
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.13.0
PyGObject==3.36.0
pygpu==0.7.6
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMuPDF==1.20.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyquaternion==0.9.9
pyrsistent==0.15.5
PySocks==1.7.1
pystow==0.5.0
PyTDC==0.4.0
pytest==7.2.0
python-apt==2.0.0+ubuntu0.20.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-docx==0.8.11
python-json-logger==2.0.7
python-Levenshtein==0.23.0
pytk==0.0.2.1
pytools==2019.1.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==18.1.1
qtconsole==5.4.1
QtPy==2.3.0
qudida==0.0.4
queuelib==1.5.0
rapidfuzz==3.4.0
rdkit-pypi==2022.9.5
regex==2023.5.5
requests==2.24.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
requestsexceptions==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-build==0.17.6
scikit-cuda==0.5.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.4.1
Scrapy==1.7.3
seaborn==0.12.2
SecretStorage==2.3.1
Send2Trash==1.8.0
sentry-sdk==1.24.0
service-identity==18.1.0
setuptools-scm==7.1.0
shapely==2.0.0
simplejson==3.19.2
six==1.15.0
sniffio==1.3.0
sos==4.2
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
ssh-import-id==5.10
sympy==1.11.1
symspellpy==6.7.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.17.1
terminaltables==3.1.10
testpath==0.4.4
Theano==1.0.4
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tk==0.1.0
tk-tools==0.16.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchsummary==1.5.1
torchvision==0.9.0+cu111
tornado==6.2
tqdm==4.50.2
traitlets==4.3.3
-e git+https://github.com/datquocnguyen/transformers.git@5be7334e3c88b3a9a2784fb19aa2e3fc2c9f327c#egg=transformers
Twisted==18.9.0
typing-extensions==3.7.4.3
tzdata==2023.3
ubuntu-advantage-tools==27.6
ufw==0.36
urllib3==1.25.11
vietocr==0.3.6
virtualenv==20.24.5
visualdl==2.4.1
w3lib==1.21.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.6
wrapt==1.12.1
xdoctest==1.1.0
xtcocotools==1.13
yapf==0.32.0
zipp==3.15.0
zope.interface==4.7.1
